# openapi.json
{"openapi":"3.1.0","info":{"title":"Langchain Vectorizing API","version":"0.1.0"},"paths":{"/{namespace}":{"post":{"summary":"Ingestdocument","operationId":"ingestDocument__namespace__post","parameters":[{"name":"namespace","in":"path","required":true,"schema":{"type":"string","title":"Namespace"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LangChainDocument"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/{namespace}":{"get":{"summary":"Searchvectorstore","operationId":"searchVectorstore_search__namespace__get","parameters":[{"name":"namespace","in":"path","required":true,"schema":{"type":"string","title":"Namespace"}},{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LangChainDocument"},"title":"Response Searchvectorstore Search  Namespace  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/document-chat/{namespace}":{"get":{"summary":"Getaichatresponse","operationId":"getAIChatResponse_document_chat__namespace__get","parameters":[{"name":"namespace","in":"path","required":true,"schema":{"type":"string","title":"Namespace"}},{"name":"question","in":"query","required":true,"schema":{"type":"string","title":"Question"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatMessage":{"properties":{"human":{"type":"boolean","title":"Human","default":false},"content":{"type":"string","title":"Content"}},"type":"object","required":["content"],"title":"ChatMessage"},"DocumentChatResponse":{"properties":{"question":{"type":"string","title":"Question"},"answer":{"type":"string","title":"Answer"},"source_documents":{"items":{"$ref":"#/components/schemas/LangChainDocument"},"type":"array","title":"Source Documents"},"chat_history":{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array","title":"Chat History"}},"type":"object","required":["question","answer","source_documents","chat_history"],"title":"DocumentChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LangChainDocument":{"properties":{"id":{"type":"null","title":"Id"},"metadata":{"type":"object","title":"Metadata","default":{}},"type":{"type":"string","title":"Type","default":"Document"},"page_content":{"type":"string","title":"Page Content"}},"type":"object","required":["id","page_content"],"title":"LangChainDocument"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}